requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
cryptography==42.0.2
qrcode==7.4.2
Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0